:sewing_needle: updated config set up
