+++ templates/config_train.yaml
@@ -1,0 +1,34 @@
+# Run name to use (if null will be replaced with <dataset_name>_<model_name>)
+run_name: null
+
+# HuggingFace model name or path to local model to fine-tune
+model_name: apple/mobilevit-xx-small
+
+# HuggingFace Dataset or path to datasets to fine-tune on
+dataset_name: pcuenq/oxford-pets
+
+# Random seed to use (e.g. for sampling the images to use)
+random_state: 42
+
+dataset_args:
+  # Name of dataset splits to train and evaluate on. If both are the same or val_split is
+  # not specified, the train_split will be randomly split into train and val splits.
+  train_split: train
+  val_split: train
+
+# Weights and Biases
+wandb:
+  entity: turing-arc
+  project: locomoset
+  log_model: end
+
+# Arguments for transformers.TrainingArguments
+training_args:
+  output_dir: output
+  overwrite_output_dir: True
+  save_strategy: "no"
+  num_train_epochs: 1
+  use_mps_device: True
+  evaluation_strategy: steps
+  eval_steps: 100
+  logging_strategy: steps
