+++ templates/metric_exp_slurm_template.sh
@@ -16,8 +16,10 @@ module load Miniconda3/4.10.3
 export CONDA_PKGS_DIRS=/tmp
 eval "$(${EBROOTMINICONDA3}/bin/conda shell.bash hook)"
 
-# This should be edited by the setup script to point in the right direction.
-CONFIGSPATH={{ config_path }}
+# This should be edited by the setup script to point in the
+# right direction for the configs and run call
+CONFIGSPATH={{ config_path }}/config_{{ config_type }}_${SLURM_ARRAY_TASK_ID}.yaml
+RUN_CALL=locomoset_run_{{ config_type }}
 
 # Define the path to your Conda environment (modify as appropriate)
 CONDA_ENV_PATH=/bask/projects/v/vjgo8416-locomoset/ARC-LoCoMoSeT/locomosetenv
@@ -25,4 +27,4 @@ conda activate ${CONDA_ENV_PATH}
 
 # Run script
 echo 'Starting task with id' ${SLURM_ARRAY_TASK_ID}
-locomoset_run_metrics ${CONFIGSPATH}/config_${SLURM_ARRAY_TASK_ID}.yaml
+$RUN_CALL ${CONFIGSPATH}
